+++ doc/sphinx/changelog.rst
@@ -33,6 +33,8 @@
 
 :ELF:
 
+  * LIEF inserted sections are not compatible with a ``strip`` after the
+    modification of the binary (see: :ref:`Adding a section/segment <format-elf-section-segment>`)
   * Relax the condition over the ``DT_SYMENT`` entry (:issue:`1177`)
   * Fix issue when parsing the dynamic table with an invalid offset (bug found
     by :github_user:`lebr0nli`)